fix(compiler): clean up latent MSVC-isms in the codegen layer
While bringing the compiler up on Apple Clang, several smaller issues
in the existing codegen surfaced. Each is a real cross-platform defect:

- `codegen_arm64.cpp` referred to `Node::slotSize()`, which does not
  exist on this branch (presumably planned for a follow-up Node-side
  refactor). The arm64 backend always uses 8-byte slots, so anchor that
  in a file-local `kArm64SlotSize` constant. The comment notes the
  fold-back path once `Node::slotSize` lands.
- `codegen_x86/tile.cpp` was using the MSVC-only `_itoa_s` four times.
  Replace with the project's existing portable `itoa()` helper from
  `stdutil`.
- `declnode.cpp` and `prognode.cpp` cast a 64-bit `Environ *` directly
  to `int` to embed it as a debugger frame handle. On 32-bit Windows
  this was a silent narrow; on 64-bit hosts the cast is rejected.
  Force the truncation through `intptr_t` so the bit pattern matches
  the legacy 32-bit Windows ABI exactly while the cast is now
  well-defined on every host. The legacy 32-bit debugger ABI is the
  only consumer.
- `exprnode.cpp` used MSVC inline assembly + `_control87` to get x87
  `fistp` rounding semantics for `FloatConstNode::intValue`. Guard the
  legacy block on `_MSC_VER && (_M_IX86 || _M_X64)` and fall back to
  `lrintf`, which on every modern host respects the IEEE
  round-to-nearest-even rounding mode that the original FPU control
  word selected.
- Replace the two backslash-style include paths in `compiler/assem.h`
  and `compiler/assem_x86/asm_insts.cpp` with forward slashes; both
  forms work on MSVC, but case-sensitive POSIX hosts choke on
  `..\linker\linker.h`.

+++ src/blitzrc/compiler/prognode.cpp
@@ -56,7 +56,8 @@ void ProgNode::translate( Codegen *g,const vector<UserFunc> &usrfuncs ){
 	if( g->debug ){
 		string t=genLabel();
 		g->s_data( "<main program>",t );
-		g->code( call( "__bbDebugEnter",local(0),iconst((int)sem_env),global(t) ) );
+		// See declnode.cpp: legacy 32-bit debugger frame opaque handle.
+		g->code( call( "__bbDebugEnter",local(0),iconst((int)(intptr_t)sem_env),global(t) ) );
 	}
 
 	//no user funcs used!
